Added encode option in video encoding
This commit introduces an "encode option" for the video encoding process in ffmpeg. The encode option has been added as a parameter in multiple functions including 'convertTsToMp4', 'RecordConfig', among others. It allows user to specify video encode options while recording a stream. The default value for this parameter is "copy". This update provides users with greater flexibility and control over the video encoding process.